requirements.txt
pyvisa
pyvisa-py
Pyserial
casperfpga
pyyaml
